Extract code related to the actual creation of the real error dataset from real.py to create_real_error_dataset.py, and organize a little bit
