+++ python/real.py
@@ -79,17 +79,6 @@ def get_real_dataset_dir(name):
 
 
 
-def structure_real_error_dataset(errors_info):
-    dataset = {}
-    for error in errors_info:
-        if error['repo'] not in dataset:
-            dataset[error['repo']] = {}
-        n_errors = len(error['errors'])
-        if n_errors not in dataset[error['repo']]:
-            dataset[error['repo']][n_errors] = []
-        dataset[error['repo']][n_errors] += [error]
-
-    return dataset
 
 
 
@@ -104,24 +93,8 @@ def structure_real_error_dataset(errors_info):
 
 
 
-def create_real_error_dataset(target):
-    if os.path.exists(target):
-        shutil.rmtree(target)
-    errors_info = load_errors_info(only_targeted=True)
-    dataset = structure_real_error_dataset(errors_info)
-    # pp.pprint(dataset)
-    for project, number_of_errors_per_file in dataset.items():
-        for number_of_errors, file_list in number_of_errors_per_file.items():
-            for id, file_info in enumerate(file_list):
-                dir = os.path.join(target, f'{project}/{number_of_errors}/{id}')
-                metadata = {
-                    'commit': file_info['commit'],
-                    'file_name': file_info['filepath'].split('/')[-1],
-                    'errors': file_info['errors']
-                }
-                create_dir(dir)
-                save_json(dir, 'metadata.json', metadata)
-                shutil.copy(file_info['filepath'], os.path.join(dir,metadata['file_name']))
+
+
 
 
 
@@ -397,9 +370,7 @@ def benchmark_stats(results):
 
 
 def main(args):
-    if len(args) >= 2 and args[1] == 'copy':
-        create_real_error_dataset(__real_dataset_dir)
-    elif len(args) >= 2 and args[1] == 'exp':
+    if len(args) >= 2 and args[1] == 'exp':
         exp(args[2:])
     elif len(args) >= 2 and args[1] == 'exp-stats':
         exp_stats(args[2:])
